change folder structure and rename
